Fix casing for libXmtpVersion across the codebase
