+++ scripts/versions.ts
@@ -6,7 +6,7 @@ interface VersionConfig {
   sdkPackage: string; // SDK package name (in node_modules/@xmtp/)
   bindingsPackage: string; // Bindings package name (in node_modules/@xmtp/)
   sdkVersion: string; // SDK version
-  bindingsVersion: string; // Bindings version
+  libXmtpVersion: string; // Bindings version
 }
 
 // Static configuration
@@ -15,31 +15,31 @@ const staticConfigs: VersionConfig[] = [
     sdkPackage: "node-sdk-mls",
     bindingsPackage: "node-bindings-mls",
     sdkVersion: "0.0.13",
-    bindingsVersion: "0.0.9",
+    libXmtpVersion: "0.0.9",
   },
   {
     sdkPackage: "node-sdk-47",
     bindingsPackage: "node-bindings-41",
     sdkVersion: "0.0.47",
-    bindingsVersion: "0.0.41",
+    libXmtpVersion: "0.0.41",
   },
   {
     sdkPackage: "node-sdk-100",
     bindingsPackage: "node-bindings-100",
     sdkVersion: "1.0.0",
-    bindingsVersion: "1.0.0",
+    libXmtpVersion: "1.0.0",
   },
   {
     sdkPackage: "node-sdk-105",
     bindingsPackage: "node-bindings-105",
     sdkVersion: "1.0.5",
-    bindingsVersion: "1.1.3",
+    libXmtpVersion: "1.1.3",
   },
   {
     sdkPackage: "node-sdk-200",
     bindingsPackage: "node-bindings-200",
     sdkVersion: "2.0.0",
-    bindingsVersion: "1.2.0-dev.bed98df",
+    libXmtpVersion: "1.2.0-dev.bed98df",
   },
 ];
